lsp, kgraph.language.server: code cleanup.
Small improvements, fixes and general cleanup, this also Fixes #40
