--- plugins/de.cau.cs.kieler.kgraph.language.server/about.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
-    "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1"/>
-<title>About</title>
-</head>
-<body lang="EN-US">
-<h2>About This Content</h2>
- 
-<p>July 01, 2015</p>	
-<h3>License</h3>
-
-<p>The Real-Time and Embedded Systems Group, Department of Computer Science, Kiel University (&quot;RTSYS&quot;)
-makes available all content in this plug-in (&quot;Content&quot;). Unless otherwise 
-indicated below, the Content is provided to you under the terms and conditions of the
-Eclipse Public License Version 1.0 (&quot;EPL&quot;).  A copy of the EPL is available 
-at <a href="http://www.eclipse.org/legal/epl-v10.html">http://www.eclipse.org/legal/epl-v10.html</a>.
-For purposes of the EPL, &quot;Program&quot; will mean the Content.</p>
-
-<p>If you did not receive this Content directly from the RTSYS, the Content is 
-being redistributed by another party (&quot;Redistributor&quot;) and different terms and conditions may
-apply to your use of any object code in the Content.  Check the Redistributor's license that was 
-provided with the Content.  If no such license exists, contact the Redistributor.  Unless otherwise
-indicated below, the terms and conditions of the EPL still apply to any source code in the Content
-and such source code may be obtained via <a href="http://www.rtsys.informatik.uni-kiel.de">http://www.rtsys.informatik.uni-kiel.de</a>.
-(Refer to &quot;KIELER&quot;, &quot;Downloads&quot;, &quot;Source Code&quot;)</p>
-
-</body>
-</html>
