+++ plugins/de.cau.cs.kieler.kgraph.language.server/build.properties
@@ -1,8 +1,5 @@
 source.. = src/,\
            xtend-gen/
 bin.includes = .,\
-               META-INF/,\
-               plugin.xml,\
-               about.html
+               META-INF/,
 bin.excludes = **/*.xtend
-src.includes = about.html
