--- plugins/de.cau.cs.kieler.kgraph.language.server/plugin.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.4"?>
-<plugin>
-   <extension
-         id="LanguageServer"
-         point="org.eclipse.core.runtime.applications">
-      <application
-            cardinality="singleton-global"
-            thread="main"
-            visible="true">
-         <run
-               class="de.cau.cs.kieler.kgraph.language.server.LanguageServer">
-         </run>
-      </application>
-   </extension>
-
-</plugin>
