+++ plugins/de.cau.cs.kieler.klighd.lsp/build.properties
@@ -1,6 +1,5 @@
 source.. = src/,\
            xtend-gen/
 bin.includes = .,\
-               META-INF/,\
-               plugin.xml
+               META-INF/
 bin.excludes = **/*.xtend
