+++ plugins/de.cau.cs.kieler.klighd.lsp/src/de/cau/cs/kieler/klighd/lsp/IActionHandler.xtend
@@ -18,7 +18,6 @@ import org.eclipse.sprotty.Action
  * Service Interface for ActionHandler.
  * 
  * @author sdo
- *
  */
 interface IActionHandler {
     /**
