+++ plugins/de.cau.cs.kieler.klighd/src/de/cau/cs/kieler/klighd/microlayout/Bounds.java
@@ -440,8 +440,8 @@ public class Bounds {
     /**
      * An immutable variant of {@link Bounds} that can be used to make sure the bounds are not
      * changed. Note that this can be ensured for usages in different packages only. Since the fields in this class are
-     * package private, any instance of these ImmutableBounds used with KLighD methods from this package may still be
-     * changed.
+     * package private, any instance of this ImmutableBounds class that is fed into KLighD methods from this package may
+     * still change the bounds, despite the naming of this class.
      *
      * @author chsch
      */
